Open eBooks Custom Login Page (#228)
* finish login

* extract home page component

* tests

* changelog

* remove file

* lint errors

* fix design
